+++ app/src/main/res/layout/fragment_map.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="campus.tech.kakao.View.MapFragment">
+    tools:context="campus.tech.kakao.View.Fragment.MapFragment">
